Debugged problem with certain keys
Keys containing an A flat had an integer value of -1 because the modulus function of Java does not handle negative numbers. To combat this, whenever I decrement a number for being flat, I take the modulus and add 12. I then modulus the sum to get the correct answer. Negative values will now be correct while positive values are unaffected.